Refactor IdPorten to be controller based
